only export what we need, mark as deprecated, move generic types up
